feat: :sparkles: add a condition to check if the task id exists in the database before trying to delete it.

+++ src/routes.js
@@ -88,6 +88,12 @@ export const routes = [
     handler: (request, response) => {
       const { id } = request.params;
 
+      const [task] = database.select("tasks", { id });
+
+      if (!task) {
+        return response.writeHead(404).end();
+      }
+
       database.delete("tasks", id);
       return response.writeHead(204).end();
     },
